Revert "/journey: résumé + life story as an opening crawl over an open-space warp cruise"
