+++ src/BadgeLink.tsx
@@ -19,9 +19,7 @@ const BadgeLink = ({ isNightMode }: { isNightMode: boolean }) => {
   // (Space3DBackground) — an invisible link over other content would
   // hijack clicks.
   const visible =
-    pathname === "/synth" ||
-    pathname === "/journey" ||
-    (pathname === "/home" && isNightMode);
+    pathname === "/synth" || (pathname === "/home" && isNightMode);
   if (!visible) return null;
 
   return (
